fix crash-and-corrupt issue when the buffer is not a multiple of WORDSIZE
Another bug in Microchip example code, whatever next?

When the programming buffer contains data, but the amount of data is not an
integer multiple of WORDSIZE, the programming function will fall into an
infinite loop. The problem with this is that the contents of the programming
buffer (generally garbage) will be continuously written to program memory,
thus completely obliterating the bootloader, CONFIG words, and anything else
it can get its sticky little paws on.

My kingdom for the zone based hardware write protection the 18F4550 had! At
least then a buggy bootloader or client app had no way of completely pooching
the boot loader.